Move NodeRegistry to openvino/pass
